% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounding_box.R
\name{bounding_box}
\alias{bounding_box}
\title{calculate the median bounding box for a yearly section}
\usage{
bounding_box(subset, img_y)
}
\arguments{
\item{subset}{subset of annotations}

\item{img_y}{image height}
}
\value{
median coordinates defining the bounding box (6 pairs)
}
\description{
Uses the corners of the yearly section as outlined by
volunteers to establish the 6 coordinate points making up
the bounding box. This is done for a subset of the full
annotated dataset.
}
