% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_linear_plots.R
\name{circular_linear_plot}
\alias{circular_linear_plot}
\title{Create species-specific linear and/or circular plot of phenology data
Linear plots show full time-series
Circular plots show annual means}
\usage{
circular_linear_plot(
  data = data,
  species_name = "Millettia laurentii",
  leg_pos = c(1, 0.2),
  title_name = "(a) evergreen"
)
}
\arguments{
\item{data}{junglerhythms data file}

\item{species_name}{list of species}

\item{title_name}{title of the plot}

\item{leg_position}{legend position (might need to change depending on how many species called)}
}
\value{
ggplot object
}
\description{
Create species-specific linear and/or circular plot of phenology data
Linear plots show full time-series
Circular plots show annual means
}
