% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_ccf_function.R
\name{climate_ccf}
\alias{climate_ccf}
\title{cross-correlations species pheno data with climate data}
\usage{
climate_ccf(
  data = data,
  climate = climate,
  climate.variable = "precip",
  species_name = "Scorodophoeus zenkeri",
  pheno = "leaf_turnover"
)
}
\arguments{
\item{data}{junglerhythms data file}

\item{climate}{yangambi climate data}

\item{climate.variable}{precip, sun or temp (one variable at a time)}

\item{species_name}{list of species}

\item{pheno}{only one phenophase}
}
\value{
dataframe significant correlation values and timing
}
\description{
cross-correlations species pheno data with climate data
}
