% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consec_years.R
\name{consecutive_timeline_sp}
\alias{consecutive_timeline_sp}
\title{for fourier analysis, no missing years can be incorporated.
If missing years within a timeline, then only the longest consecutive timeline is kept.}
\usage{
consecutive_timeline_sp(
  data = data,
  species_name = "Afzelia bipindensis",
  pheno = "leaf_turnover"
)
}
\arguments{
\item{data}{junglerhythms data file}

\item{species_name}{list of species}

\item{pheno}{only one phenophase}
}
\value{
dataframe timelines at species-level, no gaps
}
\description{
for fourier analysis, no missing years can be incorporated.
If missing years within a timeline, then only the longest consecutive timeline is kept.
}
