% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_functions.R
\name{dist_point_to_line}
\alias{dist_point_to_line}
\title{distance between a point and a line (orthogonal) (2D)}
\usage{
dist_point_to_line(coord1, coord2, coord3)
}
\arguments{
\item{coord1}{start line coordinate}

\item{coord2}{end line coordinate}

\item{coord3}{point coordinate}
}
\value{
distance
}
\description{
distance between a point and a line (orthogonal) (2D)
}
