% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_point_to_point.R
\name{dist_point_to_point}
\alias{dist_point_to_point}
\title{point to point distance (pythagoras theorem) (2D)}
\usage{
dist_point_to_point(coord1 = NULL, coord2 = NULL)
}
\arguments{
\item{coord1}{first point coordinate (x,y)}

\item{coord2}{second point coordinate (x,y)}
}
\value{
distance between points
}
\description{
point to point distance (pythagoras theorem) (2D)
}
