% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_line_sections.R
\name{extract_line_sections}
\alias{extract_line_sections}
\title{Extract line sections from annotations}
\usage{
extract_line_sections(subset, bb, img_y)
}
\arguments{
\item{subset}{subset of annotations}

\item{bb}{bounding box coordinates from bounding_box()}

\item{img_y}{image height}
}
\value{
returns a matrix of line section locations (empty 0 if nothing returned)
}
\description{
Extract line sections from annotations
}
