% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_line_circle.R
\name{intersect_line_circle}
\alias{intersect_line_circle}
\title{line circle intersection (2D)
calculates line intersections with a
circle with origin at start, only
intersections with y > start_y are
reported}
\usage{
intersect_line_circle(start, end, radius)
}
\arguments{
\item{start}{location of the center of a circle and the start of a line}

\item{end}{location of the end of a line intersecting a circle}

\item{radius}{radius of the circle to consider}
}
\value{
coordinates
}
\description{
line circle intersection (2D)
calculates line intersections with a
circle with origin at start, only
intersections with y > start_y are
reported
}
