% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_functions.R
\name{line_intersection}
\alias{line_intersection}
\title{Calculate line intersection when
provided four coordinaes defining two lines}
\usage{
line_intersection(coord1 = NULL, coord2 = NULL, coord3 = NULL,
  coord4 = NULL)
}
\arguments{
\item{coord1:}{first coordinate line 1}

\item{coord2:}{second coordinate line 1}

\item{coord3:}{first coordinate line 2}

\item{coord4:}{second coordinate line 2}
}
\description{
Calculate line intersection when
provided four coordinaes defining two lines
}
\keyword{data,}
\keyword{geometry}
\keyword{transformation,}
