% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luki_circular_linear.R
\name{luki_circular_linear_plot}
\alias{luki_circular_linear_plot}
\title{Create linear plot of phenology data}
\usage{
luki_circular_linear_plot(
  data,
  species_name = "Millettia laurentii",
  leg_pos = c(1, 0.1),
  leg_gradient = c(0, 0.2, 1),
  title_name = "(a) evergreen"
)
}
\arguments{
\item{data}{junglerhythms data file}

\item{species_name}{end coordinate fo a line}
}
\value{
ggplot object
}
\description{
Create linear plot of phenology data
}
