% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_gap_fill.R
\name{missing_years}
\alias{missing_years}
\title{In the full dataset, for each individual, obervation years without phenophase events observed
(leaf event, flowering, fruiting, seed dispersal) had been removed as missing.
Thereby, some timelines of individuals do not have consecutive dates
Here, these missing dates are filled with phenological value of NA, within the individual-specific start-end years}
\usage{
missing_years(
  data = data,
  species_name = "Afzelia bipindensis",
  pheno = "leaf_turnover",
  gapfill_missingyears = 2
)
}
\arguments{
\item{data}{junglerhythms data file}

\item{species_name}{list of species}

\item{pheno}{only one phenophase}

\item{gapfill_missingyears}{number of consecutive years with assumed observations without occured events that can be filled with zero}
}
\value{
datasets with full timelines at the individual level
}
\description{
There is an optio to fill small gaps with phenological value NA, based on the length of the missing dates:
for example, it is not unusual for these years without phenophase events to exist.
Therefore we could still assume these years as observed but without events.
To still be cautious about filling in these years, you can select the size of gap of consecutive years without phenophase events
(e.g. 2 years without observed events).
If this is longer, we keep NA
}
