% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{overview_stats}
\alias{overview_stats}
\title{Calculates summary statistics / characteristics at species-level
including number of individuals, start/end year, number of observed events,
total nr of observation years across individuals}
\usage{
overview_stats(
  data = data,
  species_name = "Afzelia bipindensis",
  pheno = "leaf_turnover"
)
}
\arguments{
\item{data}{junglerhythms data file}

\item{species_name}{list of species}

\item{pheno}{only one phenophase}
}
\value{
dataframe
}
\description{
Calculates summary statistics / characteristics at species-level
including number of individuals, start/end year, number of observed events,
total nr of observation years across individuals
}
