% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_annotations.R
\name{panel_plot}
\alias{panel_plot}
\title{Panel plot of annotations}
\usage{
panel_plot(observations, median_coord, file, image_path, output_path)
}
\arguments{
\item{observations}{annotations for a particular yearly section}

\item{median_coord}{corner point coordinates}

\item{file}{yearly section (unique filename) to process}

\item{image_path}{location of the images used in the project,
for visualization only.}

\item{output_path}{where to store images and generated data}
}
\description{
Visualizes the extracted annotations (for internal use only)
}
