% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_function_species_level.R
\name{peak_detection}
\alias{peak_detection}
\title{Peak detection and cyclicity using fourier transform}
\usage{
peak_detection(
  data = data,
  species_name = "Scorodophoeus zenkeri",
  pheno = "leaf_turnover",
  perio_plot = TRUE
)
}
\arguments{
\item{data}{junglerhythms data file}

\item{species_name}{list of species}

\item{pheno}{only one phenophase}

\item{perio_plot}{TRUE/FALSE plot periodogram}
}
\value{
timing and stats for peaks
}
\description{
Peak detection and cyclicity using fourier transform
}
