% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_dates.R
\name{phenology_dates}
\alias{phenology_dates}
\title{orthogonal projection of a point onto a line (2D)}
\usage{
phenology_dates(data, plot = TRUE,
  image_path = "~/Dropbox/Research_Projects/working/congo_phenology/data/yearly_subsets_jpg/",
  output_path = "/scratch/tmp/", internal = FALSE)
}
\arguments{
\item{data}{data to process}

\item{plot}{plot stuff}

\item{internal}{logical, save to disk or not}

\item{image_subset}{list of dimensions for image subsets}
}
\description{
orthogonal projection of a point onto a line (2D)
}
