% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_observations.R
\name{plot_raw_annotations}
\alias{plot_raw_annotations}
\title{Plot raw annotations}
\usage{
plot_raw_annotations(df)
}
\arguments{
\item{df}{annotations for a particular yearly section}
}
\description{
Visualizes the extracted annotations
}
