% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weekly_annotations.R
\name{plot_weekly_annotations}
\alias{plot_weekly_annotations}
\title{Plot weekly annotations}
\usage{
plot_weekly_annotations(data)
}
\arguments{
\item{df}{annotations for a particular yearly section}
}
\description{
Visualizes the weekly annotations
}
