% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_point_on_line.R
\name{proj_point_on_line}
\alias{proj_point_on_line}
\title{orthogonal projection of a point onto a line (2D)}
\usage{
proj_point_on_line(coord1, coord2, coord3)
}
\arguments{
\item{coord1}{start coordinate of a line}

\item{coord2}{end coordinate fo a line}

\item{coord3}{coordinates of a point}
}
\value{
coordinates x,y
}
\description{
orthogonal projection of a point onto a line (2D)
}
