% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zoo_data.R
\name{read_zoo_data}
\alias{read_zoo_data}
\title{Reads in Jungle Rhythms classification
output file (CSV) and cleans up the data
by dropping early values (trials) etc}
\usage{
read_zoo_data(file = NULL, image_file = "data/jungle_rhythms_subsets.rds",
  date = "2015-12-13", path = NULL)
}
\arguments{
\item{image_file}{file with image subset details (filename, dimensions)}

\item{path}{path where to save the data}

\item{file:}{path to a zooniverse generated CSV}

\item{date:}{date before which to discard all data (demo data)}
}
\description{
Reads in Jungle Rhythms classification
output file (CSV) and cleans up the data
by dropping early values (trials) etc
}
\keyword{data,}
\keyword{io,}
\keyword{transformation}
