% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_locations.R
\name{row_locations}
\alias{row_locations}
\title{calculate the approximate location of the rows}
\usage{
row_locations(coord)
}
\arguments{
\item{coord}{coordinates}
}
\value{
location and length of row
}
\description{
Use the corners of the yearly section (provided as a vector)
in order: bottom_left, top_left, bottom_middle,top_middle,
bottom_right, top_right
}
