% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_function_species_level.R
\docType{data}
\name{spans_lookup}
\alias{spans_lookup}
\title{Peak detection and cyclicity using fourier transform}
\format{
An object of class \code{list} of length 3.
}
\usage{
spans_lookup
}
\arguments{
\item{data}{junglerhythms data file}

\item{species_name}{list of species}

\item{pheno}{only one phenophase}

\item{perio_plot}{TRUE/FALSE plot periodogram}
}
\value{
timing and stats for peaks
}
\description{
Peak detection and cyclicity using fourier transform
}
\keyword{datasets}
