% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standlevel_phen.R
\name{standlevel_phen}
\alias{standlevel_phen}
\title{stand-level annual phenological signal
species-specific annual signals are weighted by their basal area}
\usage{
standlevel_phen(
  data = data,
  census_site = census_site,
  total_basal_area_site = total_basal_area_site,
  species_list_dorm = dorm_sp1,
  species_list_turn = turn_sp1,
  minimum_siteyears = minimum_siteyears
)
}
\arguments{
\item{data}{junglerhythms data file}

\item{census_site}{yangambi census data at site level}

\item{total_basal_area_site}{total basal area of a site}

\item{minimum_siteyears}{species not included if fewer observation-years overall (across all individuals)}

\item{species_name}{list of species}

\item{pheno}{only one phenophase}
}
\value{
dataframe
}
\description{
stand-level annual phenological signal
species-specific annual signals are weighted by their basal area
}
