% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pierlot_standlevel_phen_plotlevel.R,
%   R/standlevel_phen_plotlevel.R
\name{standlevel_phen_plotlevel}
\alias{standlevel_phen_plotlevel}
\title{plot-level annual phenological signal
species-specific annual signals are weighted by their basal area at the plot level}
\usage{
standlevel_phen_plotlevel(
  data = data,
  census_plot = census_plot,
  total_basal_area_plot = total_basal_area_plot,
  species_list_dorm = dorm_sp1,
  species_list_turn = turn_sp1,
  minimum_siteyears = minimum_siteyears
)

standlevel_phen_plotlevel(
  data = data,
  census_plot = census_plot,
  total_basal_area_plot = total_basal_area_plot,
  species_list_dorm = dorm_sp1,
  species_list_turn = turn_sp1,
  minimum_siteyears = minimum_siteyears
)
}
\arguments{
\item{data}{junglerhythms data file}

\item{census_plot}{yangambi census data at plot level}

\item{total_basal_area_plot}{total basal area of a plot}

\item{minimum_siteyears}{species not included if fewer observation-years overall (across all individuals)}

\item{species_name}{list of species}

\item{pheno}{only one phenophase}
}
\value{
dataframe

dataframe
}
\description{
plot-level annual phenological signal
species-specific annual signals are weighted by their basal area at the plot level

plot-level annual phenological signal
species-specific annual signals are weighted by their basal area at the plot level
}
