% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_annotations.R
\name{summarize_annotations}
\alias{summarize_annotations}
\title{Summarize annotations into weekly values}
\usage{
summarize_annotations(df = "data/jungle_rhythms_daily_annotations.rds",
  image_index = "data/phenology_archives_species_long_format_20190319.csv",
  plot = TRUE, internal = FALSE, output_path = "./data/")
}
\arguments{
\item{df}{junglerhythms raw annotation file}

\item{image_index}{image index data meta-data}

\item{plot}{plot output TRUE or FALSE}

\item{internal}{output results to R workspace, TRUE or FALSE}

\item{output_path}{output path where to store data}
}
\value{
ggplot object
}
\description{
weekly values assume a 48 week year
}
