% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_length.R
\name{event_length}
\alias{event_length}
\title{Calculates event length and timing}
\usage{
event_length(df)
}
\arguments{
\item{year}{year vector}

\item{week}{week vector}

\item{value}{vector with observations}
}
\value{
timing of when an event (value) switches between no event (0)
and a observation of an event (1)
}
\description{
This sub-routine will take vectors from a data frame
for a single individual and phenophase and returns the timing of
phenophases and the length of the .
}
